fix some imports, fix optional id
