+++ src/ome_types/_units.py
@@ -1,0 +1,7 @@
+import pint
+
+ureg = pint.UnitRegistry(auto_reduce_dimensions=True)
+ureg.define("reference_frame = [_reference_frame]")
+ureg.define("@alias grade = gradian")
+ureg.define("@alias astronomical_unit = ua")
+ureg.define("line = inch / 12 = li")
